added more parameters
changed upload requirements. Now requires audio file and metadata. Optional form data for whisper configuration.

added initial prompt as last transcript from talkgroup

+++ lib/config_handler.py
@@ -16,7 +16,7 @@ default_config = {
     "whisper": {
         "device": "cuda",
         "cpu_threads": 4,
-        "compute_type": "float32",
+        "compute_type": "float16",
         "model": "large-v3",
         "language": "en",
         "beam_size": 5,
